repack: add --path-walk option
Since 'git pack-objects' supports a --path-walk option, allow passing it
through in 'git repack'. This presents interesting testing opportunities for
comparing the different repacking strategies against each other.

Add the --path-walk option to the performance tests in p5313.

For the microsoft/fluentui repo [1] checked out at a specific commit [2],
the results are very interesting:

Test                                      HEAD
----------------------------------------------------------
5313.2: thin pack                         0.41(0.48+0.03)
5313.3: thin pack size                               1.2M
5313.4: thin pack with --path-walk        0.08(0.05+0.02)
5313.5: thin pack size with --path-walk             18.4K
5313.6: big pack                          4.47(30.62+0.40)
5313.7: big pack size                               19.6M
5313.8: big pack with --path-walk         6.76(9.87+0.23)
5313.9: big pack size with --path-walk              16.5M
5313.10: repack                           96.87(664.29+2.75)
5313.11: repack size                               439.5M
5313.12: repack with --path-walk          95.68(109.90+0.92)
5313.13: repack size with --path-walk              122.6M

[1] https://github.com/microsoft/fluentui
[2] e70848ebac1cd720875bccaa3026f4a9ed700e08

This repo suffers from having a lot of paths that collide in the name
hash, so examining them in groups by path leads to better deltas. Also,
in this case, the single-threaded implementation is competitive with the
full repack. This is saving time diffing files that have significant
differences from each other.

A similar, but private, repo has even more extremes during repacking:

Test                                      HEAD
-----------------------------------------------------------------
5313.10: repack                           2138.22(11961.00+17.67)
5313.11: repack size                                 6.4G
5313.12: repack with --path-walk          1351.46(1418.28+3.96)
5313.13: repack size with --path-walk              804.1M

There are small benefits in size for my copy of the Git repository:

Test                                      HEAD
-----------------------------------------------------------
5313.10: repack                           22.11(98.37+1.64)
5313.11: repack size                               126.4M
5313.12: repack with --path-walk          66.89(75.61+0.58)
5313.13: repack size with --path-walk              109.6M

As well as in the nodejs/node repository [3]:

Test                                      HEAD
-------------------------------------------------------------
5313.2: thin pack                         0.01(0.01+0.00)
5313.3: thin pack size                               1.6K
5313.4: thin pack with --path-walk        0.02(0.01+0.00)
5313.5: thin pack size with --path-walk              1.6K
5313.6: big pack                          5.35(12.43+0.32)
5313.7: big pack size                               52.2M
5313.8: big pack with --path-walk         7.12(11.97+0.27)
5313.9: big pack size with --path-walk              52.1M
5313.10: repack                           87.74(342.90+4.24)
5313.11: repack size                               739.7M
5313.12: repack with --path-walk          212.79(245.05+1.78)
5313.13: repack size with --path-walk              697.6M

[3] https://github.com/nodejs/node

This benefit also repeats in this instance in the Linux kernel repository:

Test                                      HEAD
---------------------------------------------------------------
5313.2: thin pack                         0.04(0.00+0.03)
5313.3: thin pack size                               4.6K
5313.4: thin pack with --path-walk        0.03(0.01+0.01)
5313.5: thin pack size with --path-walk              4.6K
5313.6: big pack                          21.16(62.81+1.35)
5313.7: big pack size                              158.3M
5313.8: big pack with --path-walk         36.09(55.25+0.67)
5313.9: big pack size with --path-walk             152.2M
5313.10: repack                           734.26(2149.62+31.24)
5313.11: repack size                                 2.5G
5313.12: repack with --path-walk          1457.23(1618.15+7.00)
5313.13: repack size with --path-walk                2.2G

It is important to see that even when the repository shape does not have
many name-hash collisions, there is a slight space boost to be found
using this method. Also, there is no known case where the space is
worse with --path-walk. This is of course due to the second pass where
all objects to be packed are sorted in the usual way and checked for
deltas. This second pass is usually very fast as the path-walk has
primed many objects with quality deltas that short-circuit other delta
computation attempts.

Signed-off-by: Derrick Stolee <stolee@gmail.com>